Add further clean-up (#48)
